// src/api/generated/models/LeaderboardEntryDTO.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { UsuarioDTO } from './UsuarioDTO';
export type LeaderboardEntryDTO = {
    posicion: number;
    usuario: UsuarioDTO;
    puntosTotales: number;
    aciertos: number;
};